requirements.txt
psycopg2==2.8.6
apache-airflow==2.0.1